Change to width to smallint to allow for greater values

+++ wrapper_sql.php
@@ -3,7 +3,7 @@ CREATE TABLE `wrapper` (
   `wrapper_title` varchar(255) DEFAULT NULL,
   `wrapper_url` varchar(255) NOT NULL,
   `wrapper_height` smallint(4) DEFAULT NULL,
-  `wrapper_width` tinyint(4) DEFAULT NULL,
+  `wrapper_width` smallint(4) DEFAULT NULL,
   `wrapper_scrollbars` tinyint(3) unsigned NOT NULL DEFAULT '1',
   `wrapper_userclass` tinyint(3) unsigned NOT NULL DEFAULT '0',
   PRIMARY KEY (`wrapper_id`)
